Added string interning for every string in the language.
With any new constant, we hash the string into the VMs' Hash of interned strings.

When retrieving the string, we first search the hash before creating and rehashing a new one.  This allows for fast hash searching, and findings strings without resorting to comparing all characters one by one.

We are only forced to compare when resolving collisions which is a rare operation.