Added an `add_premade_constant` function that checks for 256 instead of 257.
When emitting a global value, we do not create a new constant if the string is already interned.  That means that we would only execute add_constant.

In add_constant we added a logic check fix where counting the number of constants is being done twice, meaning that when on `make_constant` we add a new constant, the bounds check should be `256+1`.  If we only execute `add_constant` without making a new one, we do not need the `+1` fix, so for now creating a new (almost duplicate) function that solves this.

TODO -> Can refactor later though now that pressing.