Added an ActivationRecord.
An ActivationRecord contains a bytecode vector, a separate instruction pointer, and a table of local variables.  The ActivationRecord also retrieves instance and global variables using a pointer to its own bound proc.  That way variable scope is automatically managed at compile time with clear separation on syntax using `@` and `$`.  Pointers are cached to reduce dereferencing.