Refactored VM to store an activation record for main.
Main ActivationRecord is stored in the `$Margaret` object under message `main:`.

We use the API we created to first create a MargObject, create a MargMethod, and finally setting the proc field to a new MargProc.

We bind the method to its message, and the message to its object.  We retrieve the main ActivationRecord out of the primitive created, getting from each respective table.

We refactor code to work with the IP of main.