[master] - Changed the encoding of instructions to fit more bytes.
NOTE - For now we only use 16 bits as MAX_REGISTERS but technically have 29 bits available.  For 29 bits we would need to allocate the register arrays in the heap and change the implementation by using a vector potentially.  This will most likely not be the case since 1 << 16 number of registers seems like an acceptable ceiling.